Add some helpers, do some refactoring, add a schema validation test for patch member to job
